battle(sf04_fight): recover cast-free visibility test
Forward a3152d38 from the byte-exact reconstruction. The natural logical-not expression is exact under both Watcom and MSVC and restores the DOS TU placement and downstream relocation targets.

+++ src/battle.c
@@ -1732,7 +1732,7 @@ void sf04_fight(void)
         return;
     }
     do_the_fight();
-    if ((!(unsigned int)figure_list[figure_no].is_visible) & 1) {
+    if ((!figure_list[figure_no].is_visible) & 1) {
         figure_go_to_target();
         return;
     }
